refactor(admin): extract getPluginBlocks helper to lib/pluginBlocks
Copilot flagged that getPluginBlocks() was now duplicated in
components/Widgets.tsx and router.tsx with identical implementations.
Consolidate into a single shared helper under lib/pluginBlocks.ts so
the two call sites can't drift.
